+++ addons/cartel/units/CfgCommandos.hpp
@@ -1,0 +1,153 @@
+// commandos
+class GVAR(commando_g3a3) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (G3)";
+    uniformClass = "CUP_I_B_PMC_Unit_13";
+    weapons[] = { QGVAR(G3A3_modern_ris_black_burris_anpeq_flash_hider), "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { QGVAR(G3A3_modern_ris_black_burris_anpeq_flash_hider), "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_20Rnd_762x51_G3", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_Shemag_olive", "V_SmershVest_01_F", "" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_Shemag_olive", "V_SmershVest_01_F", "" };
+    backpack = "";
+};
+
+class GVAR(commando_mp5_bunkerfaust) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (AT)";
+    uniformClass = "CUP_I_B_PMC_Unit_20";
+    weapons[] = { QGVAR(MP5A5_t1), "CUP_launch_BF3", "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { QGVAR(MP5A5_t1), "CUP_launch_BF3", "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_30Rnd_9x19_MP5", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_Booniehat_mgrn", "V_CarrierRigKBT_01_light_Olive_F", "G_Bandanna_oli" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_Booniehat_mgrn", "V_CarrierRigKBT_01_light_Olive_F", "G_Bandanna_oli" };
+    icon = "iconManAT";
+};
+
+class GVAR(commando_g36_gl) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (SL)";
+    uniformClass = "CUP_I_B_PMC_Unit_42";
+    weapons[] = { QGVAR(G36A3_AG36_t1_laser), "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { QGVAR(G36A3_AG36_t1_laser), "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_30Rnd_556x45_G36", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_Booniehat_mgrn", "V_CarrierRigKBT_01_light_Olive_F", "CUP_RUS_Balaclava_blk" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_Booniehat_mgrn", "V_CarrierRigKBT_01_light_Olive_F", "CUP_RUS_Balaclava_blk" };
+    backpack = "";
+    icon = "iconManLeader";
+};
+
+class GVAR(commando_hk416) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (HK416)";
+    uniformClass = "CUP_I_B_PMC_Unit_24";
+    weapons[] = { QGVAR(HK416_CQB_Black_t1_laser), "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { QGVAR(HK416_CQB_Black_t1_laser), "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_30Rnd_556x45_PMAG_QP", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_MilCap_grn", "V_CarrierRigKBT_01_light_Olive_F", "G_Bandanna_oli" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_MilCap_grn", "V_CarrierRigKBT_01_light_Olive_F", "G_Bandanna_oli" };
+};
+
+class GVAR(commando_hk416_drum) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (AR)";
+    uniformClass = "CUP_I_B_PMC_Unit_11";
+    weapons[] = { QGVAR(HK416_Black_barska_laser), "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { QGVAR(HK416_Black_barska_laser), "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_100Rnd_TE1_Red_Tracer_556x45_BetaCMag_ar15", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "", "V_SmershVest_01_F", "CUP_RUS_Balaclava_blk" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "", "V_SmershVest_01_F", "CUP_RUS_Balaclava_blk" };
+    backpack = "";
+    icon = "IconManMG";
+};
+
+class GVAR(commando_vintorez) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (VSS)";
+    uniformClass = "CUP_I_B_PMC_Unit_13";
+    weapons[] = { QGVAR(VSSVintorez_VFG_pso1_laser), "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { QGVAR(VSSVintorez_VFG_pso1_laser), "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_20Rnd_9x39_SP5_VSS_M", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_Shemag_olive", "V_SmershVest_01_F", "" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_Shemag_olive", "V_SmershVest_01_F", "" };
+    backpack = "";
+};
+
+class GVAR(commando_ak74_rpg) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (RPG)";
+    uniformClass = "CUP_I_B_PMC_Unit_20";
+    weapons[] = { "CUP_arifle_AK74M", "CUP_launch_RPG7V", "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { "CUP_arifle_AK74M", "CUP_launch_RPG7V", "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_Booniehat_mgrn", "V_CarrierRigKBT_01_light_Olive_F", "G_Bandanna_oli" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_Booniehat_mgrn", "V_CarrierRigKBT_01_light_Olive_F", "G_Bandanna_oli" };
+    backpack = QGVAR(LAT_pack);
+    icon = "IconManAT";
+};
+
+class GVAR(commando_saiga) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (Saiga)";
+    uniformClass = "CUP_I_B_PMC_Unit_42";
+    weapons[] = { QGVAR(saiga9_1p87), "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { QGVAR(saiga9_1p87), "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_10Rnd_9x19_Saiga9", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_Booniehat_mgrn", "V_SmershVest_01_radio_F", "CUP_RUS_Balaclava_blk" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_Booniehat_mgrn", "V_SmershVest_01_radio_F", "CUP_RUS_Balaclava_blk" };
+    backpack = "";
+};
+
+class GVAR(commando_ak74) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (AK74)";
+    uniformClass = "CUP_I_B_PMC_Unit_24";
+    weapons[] = { "CUP_arifle_AK74M", "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { "CUP_arifle_AK74M", "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_30Rnd_545x39_AK74M_M", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_MilCap_grn", "V_SmershVest_01_radio_F", "G_Bandanna_oli" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_MilCap_grn", "V_SmershVest_01_radio_F", "G_Bandanna_oli" };
+    backpack = "";
+};
+
+class GVAR(commando_rpk) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Commando (RPK)";
+    uniformClass = "CUP_I_B_PMC_Unit_11";
+    weapons[] = { QGVAR(RPK74M_1p63_laser), "CUP_hgun_M9", "Put","Throw" };
+    respawnWeapons[] = { QGVAR(RPK74M_1p63_laser), "CUP_hgun_M9", "Put","Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    respawnMagazines[] = { "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_60Rnd_TE1_Green_Tracer_545x39_AK74M_M", "CUP_HandGrenade_RGD5", "CUP_HandGrenade_RGD5", "CUP_15Rnd_9x19_M9", "CUP_15Rnd_9x19_M9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "", "V_SmershVest_01_radio_F", "CUP_RUS_Balaclava_blk" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "", "V_SmershVest_01_radio_F", "CUP_RUS_Balaclava_blk" };
+    backpack = "";
+    icon = "IconManMG";
+};
