+++ addons/cartel/units/CfgCouriers.hpp
@@ -1,0 +1,30 @@
+// couriers
+class GVAR(courier_mac10) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Courier (Mac10)";
+    uniformClass = "U_C_E_LooterJacket_01_F";
+    weapons[] = { "CUP_smg_Mac10", "Put", "Throw" };
+    respawnWeapons[] = { "CUP_smg_Mac10", "Put", "Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_30Rnd_45ACP_MAC10_M", "CUP_30Rnd_45ACP_MAC10_M", "CUP_30Rnd_45ACP_MAC10_M" };
+    respawnMagazines[] = { "CUP_30Rnd_45ACP_MAC10_M", "CUP_30Rnd_45ACP_MAC10_M", "CUP_30Rnd_45ACP_MAC10_M" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_RacingHelmet_3_F", "CUP_V_CDF_OfficerBelt", "CUP_RUS_Balaclava_blk" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_RacingHelmet_3_F", "CUP_V_CDF_OfficerBelt", "CUP_RUS_Balaclava_blk" };
+    backpack = "B_Messenger_Black_F";
+};
+
+class GVAR(courier_tec9) : GVAR(Soldier_Base) {
+    scope = 2;
+    displayName = "Courier (Tec9)";
+    uniformClass = "U_C_E_LooterJacket_01_F";
+    weapons[] = { "CUP_hgun_TEC9_FA", "Put", "Throw" };
+    respawnWeapons[] = { "CUP_hgun_TEC9_FA", "Put", "Throw" };
+    items[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    respawnItems[] = { "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_packingBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage", "ACE_elasticBandage" };
+    magazines[] = { "CUP_32Rnd_9x19_TEC9", "CUP_32Rnd_9x19_TEC9", "CUP_32Rnd_9x19_TEC9" };
+    respawnMagazines[] = { "CUP_32Rnd_9x19_TEC9", "CUP_32Rnd_9x19_TEC9", "CUP_32Rnd_9x19_TEC9" };
+    linkedItems[] = { "ItemMap", "ItemCompass", "H_RacingHelmet_1_black_F", "CUP_V_CDF_OfficerBelt", "CUP_RUS_Balaclava_blk" };
+    respawnLinkedItems[] = { "ItemMap", "ItemCompass", "H_RacingHelmet_1_black_F", "CUP_V_CDF_OfficerBelt", "CUP_RUS_Balaclava_blk" };
+    backpack = "B_Messenger_Black_F";
+};
